Breaking change: rename and change default `$lpv-font-accent` to `$lpv-past-moves`
First I want to highlight I'm not convinced coloring past moves differently than the ones to come is necessary or useful, since Lichess analysis board does not have this feature and it doesn't seem to be missed. However if it's going to be implemented, better it be meaninful.

Changing the default color:

My understanding behind the idea of color past moves differently is that they are less interesting that the moves that are coming. That is why they should fade in, and the light green was confusing as it was not consistent with this idea. It also interacted poorly with good moves in the movelist, being also green.

loose conversation about it https://hq.lichess.ovh/#narrow/stream/1-general/topic/past.20move.28list.29.20colors.20of.20embed.20pgn.20viewer

Changing the scss and css variable name:

I think `$lpv-font-accent` was not a very descriptive name, since it was only used for past moves in the movelist. `$lpv-past-moves` seems better and more consistent with `$lpv-current-move`.